wip: create abstract StreamEventSourceListener class to reduce repeated code between dynamo and kinesis listeners. make dynamo listener use this class.
